+++ App.js
@@ -1,6 +1,6 @@
 import * as matrixEngine from './index.js';
 // CHANGE HERE IF YOU WANNA USE app-build.hmtl
-import {runThis} from './apps/public_3d_video_chat.js';
+import {runThis} from './apps/cube_geometry.js';
 
 var world;
 var App = matrixEngine.App;
@@ -8,20 +8,23 @@ var App = matrixEngine.App;
 if ('serviceWorker' in navigator) {
   window.addEventListener('load', function (e) {
     // navigator.serviceWorker.register('worker.js'); 
-    App.ready = true;
-    matrixEngine.Engine.initApp(webGLStart);
   });
 } else {
-  console.warn('Matrix Engine: No support for web workers in this browser.');
+  console.warn('Matrix Engine: No webWorkers for locahost OR No support for web workers in this browser.');
 }
+
+window.addEventListener('load', function (e) {
+	App.ready = true;
+	matrixEngine.Engine.initApp(webGLStart);
+});
 
 window.webGLStart = () => {
   window.App = App;
   world = matrixEngine.matrixWorld.defineworld(canvas);
   world.callReDraw();
   // Make it global for dev - for easy console/debugger access
-  window.runThis = runThis;
-  // setTimeout(() => { runThis(world); }, 1);
+  // window.runThis = runThis;
+  setTimeout(() => { runThis(world); }, 100);
   runThis(world);
 };