+++ apps/cube_geometry.js
@@ -12,22 +12,25 @@ import * as matrixEngine from "../index.js";
 export var runThis = world => {
   /* globals world App ENUMERATORS SWITCHER OSCILLATOR */
 
+	console.log('TEST APP ')
+
   var textuteImageSamplers = {
     source: ["res/images/complex_texture_1/diffuse.png"],
     mix_operation: "multiply",
   };
 
   world.Add("cubeLightTex", 1, "MyCubeTex", textuteImageSamplers);
-  App.scene.MyCubeTex.rotation.rotationSpeed.z = 70;
+ 
+  // App.scene.MyCubeTex.rotation.rotationSpeed.z = 70;
 
-  var oscilltor_variable = new matrixEngine.utility.OSCILLATOR(0.05, 2, 0.01);
+  // var oscilltor_variable = new matrixEngine.utility.OSCILLATOR(0.05, 2, 0.01);
 
   // GOOD
-  App.updateBeforeDraw.push({
-    UPDATE: () => {
-      App.scene.MyCubeTex.geometry.setScale(oscilltor_variable.UPDATE());
-    }
-  });
+  // App.updateBeforeDraw.push({
+  //   UPDATE: () => {
+  //     App.scene.MyCubeTex.geometry.setScale(oscilltor_variable.UPDATE());
+  //   }
+  // });
 
   // BAD
   // setInterval(function () {